GTK2 and backdrop state fix
GTK2: The insensitive text color for the dark theme was too similar in shade to the background color. Checkboxes in the dark theme were also pointing to the wrong asset name.

GTK3: box shadow dimensions for the window's active and backdrop state weren't consistent, leading to the transition between the two states being rather ugly